% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_exploration.R
\name{getSvd}
\alias{getSvd}
\title{A modified function to fit scalabale applications}
\usage{
getSvd(mtrx, quantile, similarity_threshold, plot_threshold = 0.05,
  varnames = NULL, plot = TRUE)
}
\description{
A modified function to fit scalabale applications
}
\author{
max
}

