% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{HClust}
\alias{HClust}
\title{Hierarchical Clustering}
\usage{
HClust(data, x, y, dmethod = "euclidean", p)
}
\arguments{
\item{data}{is the dataset containing the observations.}

\item{x}{is a reference to a column in the dataset that is could be the independent variable.}

\item{y}{is a reference to another column in the dataset that could be the dependent variable.}

\item{dmethod}{is the distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}
}
\value{
Returns a list that contains values corresponding to the cluster number, and other details
Subsets given data using column names or number and performs hierarchical clustering on the subset data with error handling.
}
\description{
Hierarchical Clustering
}
\examples{
print(HClust(iris, 'Sepal.Length', "Sepal.Width"))
}
\author{
Adithya Murali
}
\seealso{
\code{hclust}
}

