% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{HClustTable}
\alias{HClustTable}
\title{Hierarchical Clustering Classification Table}
\usage{
HClustTable(hclusters, labels, nsplit = 1)
}
\arguments{
\item{hclusters}{is a hierarchical cluster, which can be generated using the HClust function.}

\item{labels}{is a vector of values to classify the cluster data with.}

\item{nsplit}{is number of parts to split the hierarchical cluster into.}
}
\value{
Returns a table that classifies all the provided label values based on cluster number with error handling.
}
\description{
Hierarchical Clustering Classification Table
}
\examples{
print(HClustTable(HClust(iris, 'Sepal.Length', "Sepal.Width"), iris$Species, 3))
}
\author{
Adithya Murali
}
\seealso{
\code{HClust}
}

