% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{indicators}
\alias{indicators}
\title{Calculate the f1,precision,recall score of the graph}
\usage{
indicators(pred, real)
}
\arguments{
\item{pred}{Predicted graph}

\item{real}{Real graph}
}
\value{
f1,precision,recall score.
}
\description{
Calculate the f1,precision,recall score of the graph
}
\examples{
pred<-matrix(c(0,0,0,0,1,0,1,1,0),nrow=3,ncol=3)
real<-matrix(c(0,0,0,0,1,0,1,0,0),nrow=3,ncol=3)
indicators(pred,real)
}
