% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clustering.R
\name{KmeansClust}
\alias{KmeansClust}
\title{K-Means Clustering}
\usage{
KmeansClust(data, x, y, cluster, rstart = 1)
}
\arguments{
\item{data}{is the dataset containing the observations.}

\item{x}{is a reference to a column in the dataset that is could be the independent variable.}

\item{y}{is a reference to another column in the dataset that could be the dependent variable.}

\item{cluster}{is the number of clusters to perform k-means operation on.}

\item{rstart}{is how many random starting cluster assignments to try before choosing the one with the lowest within cluster variation}
}
\value{
Returns a list that contains values corresponding to the cluster number, and other details.
}
\description{
K-Means Clustering
}
\details{
Subsets given data using column names or number and performs k-means clustering on the subset data with error handling.
}
\examples{
print(KmeansClustList(iris, 'Sepal.Length', "Sepal.Width", 3))
}
\author{
Adithya Murali
}
\seealso{
\code{kmeans}
}

