% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clustering.R
\name{KmeansClustTable}
\alias{KmeansClustTable}
\title{K-Means Clustering Classification Table}
\usage{
KmeansClustTable(kclusters, labels)
}
\arguments{
\item{kclusters}{is a k-means cluster, which can be generated using the KmeansClust function.}

\item{labels}{is a vector of values to classify the cluster data with.}
}
\value{
Returns a table that classifies all the provided label values based on cluster number with error handling.
}
\description{
K-Means Clustering Classification Table
}
\examples{
print(KmeansClustTable(KmeansClust(iris, 'Sepal.Length', "Sepal.Width", 3), iris$Species))
}
\author{
Adithya Murali
}
\seealso{
\code{KmeansClust}
}

