% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMethods.R
\name{mmpcAnm}
\alias{mmpcAnm}
\title{mmpc algorithm and additive noise model}
\usage{
mmpcAnm(data)
}
\arguments{
\item{data}{The data}
}
\description{
Use the mmpc algorithm to learn a skeleton and use anm to recognize the direction
}
