% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELF.R
\docType{package}
\name{SELF-package}
\alias{SELF}
\alias{SELF-package}
\title{SELF: A Structural Equation Embedded Likelihood Framework for Causal Discovery}
\description{
Provides the SELF criteria to learn causal structure. Please cite "Ruichu Cai, Jie Qiao, Zhenjie Zhang, Zhifeng Hao. SELF: Structural Equational Embedded Likelihood Framework for Causal Discovery. AAAI,2018."
}
\author{
\strong{Maintainer}: Jie Qiao \email{qiaojie2004@vip.qq.com}

Authors:
\itemize{
  \item Ruichu Cai \email{cairuichu@gmail.com} [thesis advisor]
  \item Zhenjie Zhang \email{zhenjie@adsc.com.sg} [thesis advisor]
  \item Zhifeng Hao \email{zfhao@gdut.edu.cn} [thesis advisor]
}

}
