% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_exploration.R
\name{SVD}
\alias{SVD}
\title{SVD}
\usage{
SVD(matrix, quantile = 0.8, plot = TRUE, print = TRUE)
}
\arguments{
\item{matrix}{is the dataset containing the observations.}

\item{quantile}{is the parameter used for selecting columns that achieves the degree of explanation}
}
\value{
Returns a SVD barchar or list with vector with the column numbers that were found to be the key variables and the svd calculation
Subsets given data using column names or number and performs k-means clustering on the subset data with error handling.
}
\description{
SVD
}
\examples{
print(SVD(mtcars))
}
\author{
Adithya Murali
}
\seealso{
\code{kmeans}
}

