% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{synthetic_data_linear}
\alias{synthetic_data_linear}
\title{synthetic linear data base on the graph}
\usage{
synthetic_data_linear(G, sample_num, ratio = 1, return_noise = FALSE)
}
\arguments{
\item{G}{An adjacency matrix.}

\item{sample_num}{The number of samples}

\item{ratio}{The noise ratio It will grow or shrink the value of the noise}

\item{return_noise}{Whether return the noise of each nodes for further analysis.}
}
\value{
Return a synthetic data
}
\description{
Synthetic linear data base on the graph. The noises are sampled from the super-gaussian distribution. The coefficients are sample from U(-1,-0.5),U(0.5,1)
}
\examples{
G<-matrix(c(0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0),nrow = 4,ncol = 4)
data=synthetic_data_linear(G,100)
}
