% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{synthetic_data_nonlinear}
\alias{synthetic_data_nonlinear}
\title{synthetic nonlinear data base on the graph}
\usage{
synthetic_data_nonlinear(G, sample_num, ratio = 1, return_noise = FALSE)
}
\arguments{
\item{G}{An adjacency matrix.}

\item{sample_num}{The number of samples}

\item{ratio}{The noise ratio. It will grow or shrink the value of the noise.}

\item{return_noise}{Whether return the noise of each nodes for further analysis.}
}
\value{
Return a synthetic data
}
\description{
synthetic nonlinear data base on the graph. The data generation mechanism is y=scale(a1b1x^2+a2b2x^3+a3b3x^4+a4b4sin(x)+a5b5sin(x^2)).
}
\examples{
G<-matrix(c(0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0),nrow = 4,ncol = 4)
data=synthetic_data_nonlinear(G,100)
}
