% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMethods.R
\name{mmpcAnm}
\alias{mmpcAnm}
\title{mmpc algorithm with additive noise model}
\usage{
mmpcAnm(data)
}
\arguments{
\item{data}{The data}
}
\description{
The nonlinear data comparison algorithm. We use the mmpc algorithm to learn a causal skeleton and use ANM to recognize the direction
}
