% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{randomGraph}
\alias{randomGraph}
\title{Generate a random graph}
\usage{
randomGraph(dim, indegree)
}
\arguments{
\item{dim}{The random graph dimension}

\item{indegree}{The average indegree of random graph for each nodes}
}
\value{
Return a random graph
}
\description{
Generate a random graph based on the given dimension size and indegree
}
\examples{
randomGraph(dim=10,indegree=1)
}
